Homework 13. Auth. Add comments
