+++ app/src/main/java/ru/netology/nmedia/repository/PostRepository.kt
@@ -1,6 +1,7 @@
 package ru.netology.nmedia.repository
 
 import kotlinx.coroutines.flow.Flow
+import ru.netology.nmedia.dto.Comment
 import ru.netology.nmedia.dto.Media
 import ru.netology.nmedia.dto.MediaUpload
 import ru.netology.nmedia.dto.Post
@@ -19,6 +20,6 @@ interface PostRepository {
     suspend fun saveAsync(post: Post)
     suspend fun saveWithAttachment(post: Post, upload: MediaUpload)
     suspend fun likeByIdAsync(post: Post)
-    suspend fun getCommentsById(post: Post)
+    suspend fun getCommentsById(post: Post) : List<Comment>
     suspend fun upload(upload: MediaUpload): Media
 }
