+++ app/src/main/java/ru/netology/nmedia/repository/PostRepositoryImpl.kt
@@ -171,7 +171,7 @@ class PostRepositoryImpl(private val dao: PostDaoRoom) : PostRepository {
         }
     }
 
-    override suspend fun getCommentsById(post: Post) {
+    override suspend fun getCommentsById(post: Post) : List<Comment> {
         try {
             val response = PostsApi.retrofitService.getCommentsById(post.id)
 
@@ -181,6 +181,7 @@ class PostRepositoryImpl(private val dao: PostDaoRoom) : PostRepository {
 
             val body = response.body() ?: throw ApiError(response.code(), response.message())
             dao.insertCommentsPost(body.toEntity())
+            return body
         } catch (e: IOException) {
             throw NetworkError
         } catch (e: Exception) {
